Move Node Tests to dedicated folder
