+++ src/Travel.Tests.Unit/Workflows/Nodes/StartNodeTests.cs
@@ -10,7 +10,7 @@ using Travel.Workflows.Extensions;
 using Travel.Workflows.Nodes;
 using Travel.Workflows.Services;
 
-namespace Travel.Tests.Unit.Workflows;
+namespace Travel.Tests.Unit.Workflows.Nodes;
 
 public class StartNodeTests
 {
